+++ src/ras2fim.py
@@ -31,12 +31,16 @@ from run_ras2rem import fn_run_ras2rem
 import argparse
 import os
 import shutil
+import sys
 
 import time
 import datetime
 import fnmatch
 from re import search
 
+import shared_variables as sv
+
+# Global Variables
 b_terrain_check_only = False
 
 # $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
@@ -62,7 +66,7 @@ def fn_run_ras2fim(str_huc8_arg,
                    str_hec_path,
                    str_terrain_override,
                    run_ras2rem,
-                   str_step_override
+                   int_step
                    ):
     
     flt_start_run_ras2fim = time.time()
@@ -82,7 +86,7 @@ def fn_run_ras2fim(str_huc8_arg,
     print("  ---[v] Optional: Vertical units in: " + str(vert_unit))    
     print("  ---[t] Optional: Terrain to Utilize" + str(str_terrain_override))
     print("  ---[m] Optional: Run RAS2REM: " + str(run_ras2rem))
-    print("  ---[s] Optional: step to start at - " + str(str_step_override))
+    print("  ---[s] Optional: step to start at - " + str(int_step))
 
     print("===================================================================")
     print(" ")
@@ -92,23 +96,13 @@ def fn_run_ras2fim(str_huc8_arg,
         b_is_feet = True
     else:
         b_is_feet = False
-    if str_step_override == "None Specified - starting at the beginning":  
-        str_step_override = 0
-    int_step = int(str_step_override)
-
-    # create an output folder with checks
-    if os.path.exists(str_out_arg):
-        if os.path.exists(os.path.join(str_out_arg,'05_hecras_output','terrain_stats.csv')):
-            print(" -- ALERT: a prior sucessful run was found, delete them if you'd like to rerun ras2fim")
-            raise SystemExit(0)
-        elif int_step==0:
-            print(" -- ALERT: a prior partially sucessful run was found, deleteing and retrying this.")
-            shutil.rmtree(str_out_arg, ignore_errors=False, onerror=None)
-    os.mkdir(str_out_arg)   
     
     # ---- Step 1: create_shapes_from_hecras ----
     # create a folder for the shapefiles from hec-ras
-    str_shapes_from_hecras_dir = os.path.join(str_out_arg, "01_shapes_from_hecras") 
+    print()
+    print ("+++++++ Processing for code STEP 1 +++++++" )
+
+    str_shapes_from_hecras_dir = os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_SHAPES_FROM_HECRAS) 
     if not os.path.exists(str_shapes_from_hecras_dir):
         os.mkdir(str_shapes_from_hecras_dir)
     
@@ -121,7 +115,10 @@ def fn_run_ras2fim(str_huc8_arg,
 
     # ------ Step 2: conflate_hecras_to_nwm -----    
     # do whatever is needed to create folders and determine variables
-    str_shapes_from_conflation_dir = os.path.join(str_out_arg, "02_shapes_from_conflation")
+    print()
+    print ("+++++++ Processing for code  STEP 2 +++++++" )
+
+    str_shapes_from_conflation_dir = os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_SHAPES_FROM_CONF)
     if not os.path.exists(str_shapes_from_conflation_dir):
         os.mkdir(str_shapes_from_conflation_dir)
     
@@ -136,9 +133,11 @@ def fn_run_ras2fim(str_huc8_arg,
     # ------ Step 3: get_usgs_dem_from_shape or clip_dem_from_shape ----    
     str_area_shp_name = str_huc8_arg + "_huc_12_ar.shp"
     str_input_path = os.path.join(str_shapes_from_conflation_dir, str_area_shp_name)
-    
+    print()
+    print ("+++++++ Processing for code  STEP 3 +++++++" )
+
     # create output folder
-    str_terrain_from_usgs_dir = os.path.join(str_out_arg, "03_terrain")
+    str_terrain_from_usgs_dir = os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_TERRAIN)
     if not os.path.exists(str_terrain_from_usgs_dir):
         os.mkdir(str_terrain_from_usgs_dir)
         
@@ -174,15 +173,16 @@ def fn_run_ras2fim(str_huc8_arg,
     # -------------------------------------------
 
     # ------ Step 4: convert_tif_to_ras_hdf5 ----- 
-    # 
-    
+      
     # folder of tifs created in third script (get_usgs_dem_from_shape)
     # str_terrain_from_usgs_dir
     
     # create a converted terrain folder
-    str_hecras_terrain_dir = os.path.join(str_out_arg, "04_hecras_terrain")
+    str_hecras_terrain_dir = os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_HECRAS_TERRAIN)
     if not os.path.exists(str_hecras_terrain_dir):
         os.mkdir(str_hecras_terrain_dir)
+    print()
+    print ("+++++++  Processing for code STEP 4 +++++++" )
         
     str_area_prj_name = str_huc8_arg + "_huc_12_ar.prj"
     str_projection_path = os.path.join(str_shapes_from_conflation_dir, str_area_prj_name)
@@ -201,9 +201,11 @@ def fn_run_ras2fim(str_huc8_arg,
     # str_terrain_from_usgs_dir
     
     # create a converted terrain folder
-    str_fim_out_dir = os.path.join(str_out_arg, "05_hecras_output")
-    if not os.path.exists(str_fim_out_dir):
-        os.mkdir(str_fim_out_dir)
+    str_hecras_out_dir = os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_HECRAS_OUTPUT)
+    if not os.path.exists(str_hecras_out_dir):
+        os.mkdir(str_hecras_out_dir)
+    print()
+    print ("+++++++ Processing for code  STEP 5 +++++++" )
     
     # path to standard input (PlanStandardText01.txt, PlanStandardText02.txt, ProjectStandardText01.txt )
     str_std_input_path = os.getcwd() # assumed as in directory executing script
@@ -221,7 +223,7 @@ def fn_run_ras2fim(str_huc8_arg,
     if int_step <= 5:
         fn_create_fim_rasters(str_huc8_arg,
                               str_shapes_from_conflation_dir,
-                              str_fim_out_dir,
+                              str_hecras_out_dir,
                               str_projection_path,
                               str_hecras_terrain_dir,
                               str_std_input_path,
@@ -236,23 +238,28 @@ def fn_run_ras2fim(str_huc8_arg,
     str_output_crs = "EPSG:3857"
     # ==============================================
     
+    print()
+    print ("+++++++  Processing for code STEP 6 +++++++" )
     if int_step <= 6:
-        fn_simplify_fim_rasters(str_fim_out_dir,
+        fn_simplify_fim_rasters(str_hecras_out_dir,
                                 flt_resolution_depth_grid,
                                 str_output_crs)
     # ----------------------------------------
     
 
     # ------ Step 7: calculate terrain statistics -----
+    print()
+    print ("+++++++ Processing for code  STEP 7 +++++++" )
+
     if int_step <= 7:
-        fn_calculate_all_terrain_stats(str_fim_out_dir)
+        fn_calculate_all_terrain_stats(str_hecras_out_dir)
     # -------------------------------------------------
 
     # ------ Step 8: run ras2rem -----
+    print()
+    print ("+++++++ Processing for code  STEP 8 +++++++" )
     if int_step <= 8 and run_ras2rem:
-        # define output folder
-        str_ras2rem_dir = os.path.join(str_out_arg, "06_ras2rem")
-        fn_run_ras2rem(str_fim_out_dir, str_ras2rem_dir)
+        fn_run_ras2rem(str_out_arg)
 
     flt_end_run_ras2fim = time.time()
     flt_time_pass_ras2fim = (flt_end_run_ras2fim - flt_start_run_ras2fim) // 1
@@ -260,105 +267,121 @@ def fn_run_ras2fim(str_huc8_arg,
     
     print('Total Compute Time: ' + str(time_pass_ras2fim))
     
+
+def create_input_args_log (**kwargs):
+
+    '''
+    Overview:
+        This method takes all incoming arguments, cycles through them and put them in a file
+    Inputs:
+        **kwargs is any dictionary of key / value pairs
+    '''
+
+    r2f_huc_output_dir = kwargs.get("r2f_huc_output_dir")
+
+    arg_log_file = os.path.join(r2f_huc_output_dir, "run_arguments.txt")
     
-    
-# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^    
+    # Remove it if is aleady exists (relavent if we add an override system)
+    if (os.path.exists(arg_log_file)):
+        os.remove(arg_log_file)
+
+    # start with the processing date
+    utc_now = datetime.datetime.now(datetime.timezone.utc)
+    str_date = utc_now.strftime("%Y-%m-%d")
+
+    # The file can be parsed later by using the two colons and the line break if ever required
+    # We are already talking about using data in this file for metadata files
+    # especially as the DEM's becomed versions in the input files which meta data
+    # will need to know what fim version of the DEM was used.
+    with open(arg_log_file, "w") as arg_file:
+        arg_file.write(f"process_date == {str_date}\n")
+        arg_file.write(f"command_line_submitted == {(' '.join(sys.argv))}\n")
+
+        for key, value in kwargs.items():
+            arg_file.write("%s == %s\n" % (key, value))
 
 
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='========== RUN RAS2FIM FOR A HEC-RAS 1-D DATASET (HUC8) ==========')
-    
-    parser.add_argument('-w',
-                        dest = "str_huc8_arg",
-                        help='REQUIRED: HUC-8 watershed that is being evaluated: Example: 10170204',
-                        required=True,
-                        metavar='STRING',
-                        type=str)
-    
-    parser.add_argument('-i',
-                        dest = "str_ras_path_arg",
-                        help=r'REQUIRED: path containing the HEC-RAS files: Example C:\HEC\ras_folder',
-                        required=True,
-                        metavar='DIR',
-                        type=str)
-
-    parser.add_argument('-o',
-                        dest = "str_out_arg",
-                        help=r'REQUIRED: path to write all the outputs: Example C:\HEC\output_folder',
-                        required=True,
-                        metavar='DIR',
-                        type=str)
-    
-    parser.add_argument('-p',
-                        dest = "str_crs_arg",
-                        help=r'REQUIRED: projection of HEC-RAS models: Example EPSG:26915',
-                        required=True,
-                        metavar='STRING',
-                        type=str)
-    
-    parser.add_argument('-n',
-                        dest = "str_nation_arg",
-                        help=r'REQUIRED: path to national datasets: Example: E:\X-NWS\X-National_Datasets',
-                        required=True,
-                        metavar='DIR',
-                        type=str)
-    
-    parser.add_argument('-r',
-                        dest = "str_hec_path",
-                        help=r'REQUIRED: path to HEC-RAS 6.0: Example: "C:\Program Files (x86)\HEC\HEC-RAS\6.0" (wrap in quotes)',
-                        required=True,
-                        metavar='DIR',
-                        type=str)
-    
-    parser.add_argument('-v',
-                        dest = "vert_unit",
-                        help='OPTIONAL: define the vertical units of the project, one of "meter" or "foot", leave black to have ras2fim check automatically',
-                        required=False,
-                        default='check',
-                        metavar='STRING',
-                        type=str)
-
-    parser.add_argument('-t',
-                        dest = "str_terrain_override",
-                        help=r'OPTIONAL: path to DEM/VRT to use for mapping: Example: G:\x-fathom\vrt\texas_fathom_20211001.vrt',
-                        required=False,
-                        default='None Specified - using USGS WCS',
-                        metavar='PATH',
-                        type=str)
-
-    parser.add_argument('-m',
-                        dest="rem_outputs",
-                        help=r'OPTIONAL: flag to dictate RAS2REM execution: Enter false to skip, defaults to TRUE',
-                        required=False,
-                        default=True,
-                        metavar='T/F',
-                        type=str2bool)
-
-    parser.add_argument('-s',
-                        dest = "str_step_override",
-                        help=r'OPTIONAL: step of processing to start on',
-                        required=False,
-                        default='None Specified - starting at the beginning',
-                        metavar='STRING',
-                        type=str)
+def init_and_run_ras2fim(str_huc8_arg, 
+                         str_crs_arg,
+                         r2f_huc_output_dir,
+                         str_hec_path  = sv.DEFAULT_HECRAS_ENGINE_PATH,
+                         str_ras_path_arg = sv.HECRAS_INPUT_DEFAULT_OWP_RAS_MODELS,
+                         str_nation_arg  = sv.INPUT_DEFAULT_X_NATIONAL_DS_DIR,
+                         vert_unit = 'check',
+                         str_terrain_override = 'None Specified - using USGS WCS',
+                         rem_outputs = True,
+                         str_step_override = 'None Specified - starting at the beginning'):
 
 
-    
-    args = vars(parser.parse_args())
-    
-    str_huc8_arg = args['str_huc8_arg']
-    str_ras_path_arg = args['str_ras_path_arg']
-    str_out_arg = args['str_out_arg']
-    str_crs_arg = args['str_crs_arg']
-    vert_unit = args['vert_unit']
-    str_nation_arg = args['str_nation_arg']
-    str_hec_path = args['str_hec_path']
-    str_terrain_override = args['str_terrain_override']
-    rem_outputs = args['rem_outputs']
-    str_step_override = args['str_step_override']
+    ####################################################################
+    ####  Some validation of input, but mostly setting up pathing ######
 
-    
+    # -w   (ie 12090301)
+    if (len(str_huc8_arg) != 8):
+        raise ValueError("the -w flag (HUC8) is not 8 characters long")
+    if (str_huc8_arg.isnumeric() == False): # can handle leading zeros
+        raise ValueError("the -w flag (HUC8) does not appear to be a HUC8")
 
+    # -i  (ie OWP_ras_models\models) (HECRAS models)
+    if (os.path.exists(str_ras_path_arg) == False) and (str_ras_path_arg != sv.HECRAS_INPUT_DEFAULT_OWP_RAS_MODELS):
+        raise ValueError("the -i arg (ras path arg) does not appear to be a valid folder.")
+        
+    # -o  (ie 12090301_meters_2277_test_1) or some full custom path
+    # We need to remove the the last folder name and validate that the parent paths are valid
+    is_invalid_path = False
+    if ("\\" in r2f_huc_output_dir):  # submitted a full path
+        parent_dir_name = os.path.dirname(r2f_huc_output_dir)
+        if (os.path.exists(parent_dir_name) == False): # parent path must exist
+            is_invalid_path = True
+    else: # they provide just a child folder (base path name)
+        r2f_huc_output_dir = os.path.join(sv.R2F_DEFAULT_OUTPUT_MODELS, r2f_huc_output_dir)
+        if (os.path.exists(r2f_huc_output_dir) == True): # child folder must not exist
+            is_invalid_path = True
+
+    if (is_invalid_path == True):
+        raise ValueError('The -o child folder (base path) can not pre-exist as we will create it, but the ' \
+                         ' parent folders must exist. See code notes in the __main__ section for details and examples')
+
+
+    # TODO: step system not fully working and needs to be fixed.
+    # create an output folder with checks
+    #if os.path.exists(str_out_arg):
+    #    if os.path.exists(os.path.join(str_out_arg, sv.R2F_OUTPUT_DIR_HECRAS_OUTPUT, 'terrain_stats.csv')):
+    #        print(" -- ALERT: a prior sucessful run was found, delete them if you'd like to rerun ras2fim")
+    #        raise SystemExit(0)
+    #    elif int_step==0:
+    #        print(" -- ALERT: a prior partially sucessful run was found, deleteing and retrying this.")
+    #        shutil.rmtree(str_out_arg, ignore_errors=False, onerror=None)
+    #else:
+    os.mkdir(r2f_huc_output_dir) # pathing has already been validated and ensure the child folder does not pre-exist
+
+
+    # -n  (ie: inputs\\X-National_Datasets)
+    if (os.path.exists(str_nation_arg) == False) and (str_nation_arg != sv.INPUT_DEFAULT_X_NATIONAL_DS_DIR):
+        raise ValueError("the -n arg (inputs x national datasets path arg) does not appear to be a valid folder.")
+
+
+    # -r  (ie: C:\Program Files (x86)\HEC\HEC-RAS\6.0)
+    if (os.path.exists(str_hec_path) == False):
+        raise ValueError("the -r arg (HEC-RAS engine path) does not appear to be correct.")
+
+    # -t  (ie: blank or a path such as inputs\12090301_dem_meters_0_2277.tif)
+    if (str_terrain_override != "None Specified - using USGS WCS"):
+        if (os.path.exists(str_terrain_override) == False): # might be a full path 
+            raise ValueError("the -t arg (terrain override) does not appear to be correct a valid path and file.")
+
+
+    if str_step_override == "None Specified - starting at the beginning":  
+        int_step = 0
+    else:
+        int_step = int(str_step_override)
+
+
+    # Save incoming args and a few new derived variables created to this point into a log file
+    # Careful... when **locals() is called, it will include ALL variables in this function to this point.
+    create_input_args_log(**locals())
+
+    # Setup enviroment logic
     if vert_unit == 'check':
         matches = []
         for root, dirnames, filenames in os.walk(str_ras_path_arg):
@@ -395,14 +418,135 @@ if __name__ == '__main__':
             elif identical_array[0] == "foot":
                 vert_unit = "foot"
 
+
     fn_run_ras2fim(str_huc8_arg,
                    str_ras_path_arg,
-                   str_out_arg,
+                   r2f_huc_output_dir,
                    str_crs_arg,
                    vert_unit,
                    str_nation_arg,
                    str_hec_path,
                    str_terrain_override,
                    rem_outputs,
-                   str_step_override,
-                   )
+                   int_step )
+
+    
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^    
+
+if __name__ == '__main__':
+
+
+    # Sample usage:
+    # Using all defaults:
+    #     python ras2fim.py -w 12090301 -o 12090301_meters_2277_test_2 -p EPSG:2277 -t C:\ras2fim_data\inputs\12090301_dem_meters_0_2277.tif
+
+    # There are a number of ways to use arguments:
+    #
+    # 1) You can use the defaults, which will use the folder structure as seen in the docs/INSTALL.md. It will require you to 
+    #    to create folders in a specific pattern and specific names.
+
+    # 2) You can also use the current code in legacy fashion with all of it's previous arguments.
+    #    ie) python ras2fim.py -w 10170204 -i C:\HEC\input_folder -o C:\HEC\output_folder -p EPSG:26915
+    #            -v False -n E:\X-NWS\X-National_Datasets -r "C:\Program Files (x86)\HEC\HEC-RAS\6.0"
+    #        
+    #    But any and all optional arguments can be overridden, so let's try this version:
+    #    ie) python ras2fim.py -w 12090301 -i C:\HEC\input_folder -o c:/users/my_user/desktop/ras2fim_outputs/12090301_meters_2277_test_2 -p EPSG:2277
+    #                          -t C:\ras2fim_data\inputs\12090301_dem_meters_0_2277.tif -n E:\X-NWS\X-National_Datasets
+    #    
+    #        - When the -n arg not being set, it defaults to c:/ras2fim_data/inputs/X-National_datasets.
+    #
+    #        - The -o arg is required, but can be either a full path (as shown above), or a simple folder name. 
+    #             ie) -o c:/users/my_user/desktop/ras2fim_outputs/12090301_meters_2277_test_2  (the 6 folders of 01, 02... 06 will be under this folder can not be overriden).
+    #                 OR
+    #                 -o 12090301_meters_2277_test_3  (We will use the root default pathing and become c:/ras2fim_data/outputs_ras2fim_models/12090301_meters_2277_test_3)
+    #             As before.. if the final folder of 12090301_meters_2277_test_2 can not pre-exist and we will create it.
+    #             However, as before, it will error out if the parent path (ie c:/users/my_user/desktop/ras2fim_outputs) does not exist.
+    #
+    #         - The -i arg is optional and defaults to c:/ras2fim_data/OWP_ras_models/models. Each file or subfolder in this directory will be used as input into the ras2fim.py code.
+    #              Again, you an override this to any location you like.
+    #
+    #  Note: Careful on copy / pasting commands directly from here as some have line breaks for display purposes.
+    #        Python command line commands don't like line breaks and you will need to remove them.
+
+
+    parser = argparse.ArgumentParser(description='========== RUN RAS2FIM FOR A HEC-RAS 1-D DATASET (HUC8) ==========')
+
+    parser.add_argument('-w',
+                        dest = "str_huc8_arg",
+                        help = 'REQUIRED: HUC-8 watershed that is being evaluated: Example: 10170204',
+                        required = True,
+                        type = str)  # has to be string so it doesn't strip the leading zero
+
+    parser.add_argument('-p',
+                        dest = "str_crs_arg",
+                        help = r'REQUIRED: projection of HEC-RAS models: Example EPSG:2277',
+                        required = True,
+                        type = str)
+
+    parser.add_argument('-o',
+                        dest = "r2f_huc_output_dir",
+                        help = r'REQUIRED: This can be used in one of two ways. You can submit either a full path' \
+                               r' such as c:\users\my_user\Desktop\myoutput OR you can add a simple final folder name.' \
+                                ' Please see the embedded notes in the __main__ section of the code for details and  examples.',
+                        required = True,
+                        type = str) 
+    
+    parser.add_argument('-r',
+                        dest = "str_hec_path",
+                        help = r'OPTIONAL: path to HEC-RAS 6.0: Defaults to C:\Program Files (x86)\HEC\HEC-RAS\6.0' \
+                               r' but you can override it, Example: "C:\Program Files (x86)\HEC\HEC-RAS\6.3" (wrap in quotes)',
+                        required = False,
+                        default = sv.DEFAULT_HECRAS_ENGINE_PATH,
+                        type = str)
+
+    parser.add_argument('-i',
+                        dest = "str_ras_path_arg",
+                        help = r'OPTIONAL: path containing the HEC_RAS files: Example -i C:\HEC\input_folder\my_models.' \
+                               r' Defaults to c:\ras2fim_datas\OWP_ras_models\models.',
+                        default = sv.HECRAS_INPUT_DEFAULT_OWP_RAS_MODELS,
+                        required = False,
+                        type = str)
+
+    parser.add_argument('-n',
+                        dest = "str_nation_arg",
+                        help = r'OPTIONAL: path to national datasets: Example: E:\X-NWS\X-National_Datasets.' \
+                               r' Defaults to c:\ras2fim_data\inputs\X-National_Datasets.',
+                        default = sv.INPUT_DEFAULT_X_NATIONAL_DS_DIR,
+                        required = False,
+                        type = str)
+        
+    parser.add_argument('-v',
+                        dest = "vert_unit",
+                        help='OPTIONAL: define the vertical units of the project, one of "meter" or "foot",' \
+                             ' leave black to have ras2fim check automatically (Default = check).',
+                        required = False,
+                        default = 'check',
+                        type = str)
+
+    parser.add_argument('-t',
+                        dest = "str_terrain_override",
+                        help = r'OPTIONAL: full path to DEM Tif to use for mapping: Example: c:\ras2fim_data\inputs\some_dem.tif.' \
+                               r'Defaults to calling USGS website, but note.. it can be unstable and throw 404 and 500 errors.',
+                        required = False,
+                        default = 'None Specified - using USGS WCS',
+                        type = str)
+
+    parser.add_argument('-m',
+                        dest = "rem_outputs",
+                        help = r'OPTIONAL: flag to dictate including RAS2REM execution: Enter false to skip, defaults to TRUE.',
+                        required = False,
+                        default = True,
+                        metavar = 'T/F',
+                        type = str2bool)
+
+    parser.add_argument('-s',
+                        dest = "str_step_override",
+                        help = r'OPTIONAL: step of processing to start on. Note: This feature is temporarily not working.',
+                        required = False,
+                        default = 'None Specified - starting at the beginning',
+                        type = str)
+
+    
+    args = vars(parser.parse_args())
+    
+    init_and_run_ras2fim(**args)